chore(protocol): re-org mod visibilities and errors
